+++ src/spiral.rs
@@ -1,27 +1,27 @@
 
-pub fn next_spiral(n: i64) -> Vec<i32> {    
-    let r: i32 = ((((n+1) as f64).sqrt() - 1.0) / 2.0).floor() as i32 + 1;
-    let p: i32 = (8 * (r as i32) * ((r as i32) - 1)) / 2;
-    let en: i32 = (r as i32) * 2;
-    let a: i32 = (1 + n + (p as i64)) as i32 % ((r as i32) * 8);
+pub fn next_spiral(n: i64) -> Vec<i64> {    
+    let r: i64 = ((((n+1) as f64).sqrt() - 1.0) / 2.0).floor() as i64 + 1;
+    let p: i64 = (8 * (r) * ((r) - 1)) / 2;
+    let en: i64 = (r) * 2;
+    let a: i64 = (1 + n + (p)) % ((r) * 8);
     
-    let mut pos: Vec<i32> = ([0, 0]).to_vec();
-    match a as i64 / (r * 2) as i64{
+    let mut pos: Vec<i64> = ([0, 0]).to_vec();
+    match a / (r * 2) {
         0=> { 
-            pos[0] = a - (r as i32);
-            pos[1] = -(r as i32);
+            pos[0] = a - (r as i64);
+            pos[1] = -(r);
         }
         1=> { 
-            pos[0] = r as i32;
-            pos[1] = (a % en) - (r as i32);
+            pos[0] = r;
+            pos[1] = (a % en) - (r);
         }
         2=> {
-            pos[0] =  (r as i32) - (a % en);
-            pos[1] = r as i32;
+            pos[0] =  (r) - (a % en);
+            pos[1] = r;
         }
         3=> { 
-            pos[0] = -(r as i32);
-            pos[1] = (r as i32) - (a % en);
+            pos[0] = -(r);
+            pos[1] = (r) - (a % en);
         }
         _=> {}
     }
